+++ src/main/java/com/project/phonecaseshop/entity/dto/memberDto/MemberResponseDto.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberResponseDto {
-    private Long memberId;
+    private int memberId;
     private String memberEmail;
     private String memberPassword;
     private String memberNickname;
-    private Long memberPoint;
+    private int memberPoint;
     private String memberAddress;
     private String memberDetailAddress;
     private String memberStatus;
